+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Loadouts with a complete set of equipped armor now include a stat bar that will tell you the stat tiers of the equipped loadout pieces.
 * Loadouts with non-equipping items now won't *de-equip* those items if they're already equipped. #1567
 * The count of items in your loadout is now more accurate.
+* DIM is now better at figuring out which platforms you have Destiny accounts on.
+* DIM is faster!
 
 # v3.17.1